+++ packages/plugins/cattle/src/plugin.ts
@@ -1,8 +1,0 @@
-import { cattle } from "./commands/cattle.js";
-
-export const plugin = {
-  name: "cattle",
-  command: cattle,
-};
-
-export const command = cattle;